Move log logic to logging module
